requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
python-multipart
python-dotenv
google